// src/app/models/kvartalniIzvestaj.ts
import { starosneKategorije } from './starosneKategorije';

export class kvartalniIzvestaj{
    poseteUkupno: starosneKategorije;
    prvePosete: starosneKategorije;
    serijskePosete: starosneKategorije;
    plombiraniZubiBezLecenja: starosneKategorije;
    plombiraniZubiSaLecenjem: starosneKategorije;
    hirurskeIntervencijeVadjenje: starosneKategorije;
    hirurskeIntervencijeOstalo: starosneKategorije;
    pokretneProtezeTotalne: starosneKategorije;
    pokretneProtezeParcijalne: starosneKategorije;
    fiksneProtezeKrunice: starosneKategorije;
    fiksneProtezeClanovi: starosneKategorije;
    ortodoncijaPrviPregledi: starosneKategorije;
    ortodoncijaPokretniAparati: starosneKategorije;
    ortodoncijaTerapeutskeIntervencije: starosneKategorije;
    ortodoncijaFiksniAparati: starosneKategorije;
    mekaTkivaUsneSupljine: starosneKategorije;
}